Task 2 Complete
Both Jada and I didn't know much about chess, so it took us some time and many websites to understand the purpose and endgame of a chess game. Not to mention, we had to figure out the different aspects and how to move and progress them correctly throughout the board.