Task #11 Complete
Since, Jada has been working more on the coding aspect than I have I decided to focus on the powerpoint presentation today. I created the powerpoint for both assignments Sprint #3 and the final presentation, with this organization strategies already defined and created it should help us only focus on the coding aspect and making sure that the application can run for the overall project more.