Task #1 Complete
This task wasn't as difficult to complete, but it did help us have somewhere to start. Jada, who has been working on the code so diligently, created the title for the project on Visual Studio Code, along with additional files. I have completed the title page on the final presentation powerpoint that we will be using towards the end of the semester.