Task #5 Complete
For this task we decided that I would assist Jada in completing the code aspect of the assignment and she would make sure it would run properly. Not to mention, we will both be present to fix any problems that are present in the project. I will be completing the final powerpoint, since Jada will be doing majority of the coding and I will be more assisting where I can in the coding aspect.