Task #4 Complete
Me and Jada have decided that we will communicate using Teams since those are the quickest way we are able to contact one another. Not to mention, we will join a team meeting or meet in person if it is needed if we are physical problems with Visual Studio Code or any other physical problems that we need to meet face-to-face about.